# docs/openapi.v1.yaml
openapi: 3.0.3
info:
  title: MADAVOLA API
  version: v1
  description: API contract v1 (spec-first). Certains endpoints sont planifiés.
servers:
  - url: /api/v1
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
...
  /actors:
    post:
      tags: [actors]
      summary: Create actor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActorCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
    get:
      tags: [actors]
      summary: List actors
      parameters:
        - in: query
          name: role
          schema: { type: string }
        - in: query
          name: commune_code
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Actor"
  /actors/{actor_id}:
    get:
      tags: [actors]
      summary: Get actor
      parameters:
        - in: path
          name: actor_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
    patch:
      tags: [actors]
      summary: Update actor (sensible)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActorUpdate"
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    AuthTokens:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
    LoginRequest:
      type: object
      properties:
        identifier: { type: string }
        password: { type: string }
      required: [identifier, password]
    RefreshRequest:
      type: object
      properties:
        refresh_token: { type: string }
      required: [refresh_token]
    Actor:
      type: object
      properties:
        id: { type: string }
        type_personne: { type: string }
        nom: { type: string }
        prenoms: { type: string }
        cin: { type: string }
        nif: { type: string }
        stat: { type: string }
        rccm: { type: string }
        telephone: { type: string }
        email: { type: string }
        region_code: { type: string }
        district_code: { type: string }
        commune_code: { type: string }
        fokontany_code: { type: string }
        status: { type: string }
    ActorCreate:
      type: object
      properties:
        type_personne: { type: string }
        nom: { type: string }
        prenoms: { type: string }
        cin: { type: string }
        nif: { type: string }
        stat: { type: string }
        rccm: { type: string }
        telephone: { type: string }
        password: { type: string }
        region_code: { type: string }
        district_code: { type: string }
        commune_code: { type: string }
        fokontany_code: { type: string }
        geo_point_id: { type: integer }
        roles:
          type: array
          items: { type: string }
      required: [type_personne, nom, telephone, password, commune_code, geo_point_id, roles]
    ActorUpdate:
      type: object
      properties:
        status: { type: string }
        justification: { type: string }
...